Reorganize and rename county end of round components.
