+++ client/src/component/county/Audit/Container.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
 
-import EndOfRoundPageContainer from './EndOfRoundPageContainer';
+import EndOfRoundPageContainer from './EndOfRound/PageContainer';
 import CountyAuditPage from './Page';
 
 import notice from 'corla/notice';
